@@ -14,37 +14,40 @@
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
     "@reduxjs/toolkit": "^1.6.2",
-    "antd": "^4.16.13",
-    "axios": "^0.21.4",
-    "date-fns": "^2.26.0",
+    "antd": "^4.17.2",
+    "axios": "^0.24.0",
+    "component-emitter": "^1.3.0",
+    "date-fns": "^2.27.0",
     "md5": "^2.3.0",
     "mockjs": "^1.1.0",
     "noble-ed25519": "^1.2.6",
     "nodejs-base64": "^2.0.0",
     "oicq": "^1.21.2",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-redux": "^7.2.5",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "sass": "^1.42.1",
-    "socket.io-client": "^4.2.0"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "react-router": "^6.0.2",
+    "react-router-dom": "^6.0.2",
+    "sass": "^1.44.0",
+    "socket.io-client": "^4.4.0"
   },
   "devDependencies": {
+    "@types/component-emitter": "^1.2.11",
     "@types/md5": "^2.3.1",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "@vitejs/plugin-react": "^1.0.0",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-redux": "^7.1.20",
+    "@types/react-router-dom": "^5.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.5.0",
+    "@typescript-eslint/parser": "^5.5.0",
+    "@vitejs/plugin-react": "^1.1.0",
     "eslint": "^7.32.0",
-    "eslint-config-alloy": "^4.3.0",
-    "eslint-plugin-react": "^7.26.1",
-    "prettier": "^2.4.1",
-    "typescript": "^4.4.3",
-    "vite": "^2.6.0",
+    "eslint-config-alloy": "^4.4.0",
+    "eslint-plugin-react": "^7.27.1",
+    "prettier": "^2.5.0",
+    "typescript": "^4.5.2",
+    "vite": "^2.6.14",
     "vite-plugin-mock": "^2.9.6"
   },
-  "packageManager": "yarn@3.0.2"
+  "packageManager": "yarn@3.1.1"
 }
