@@ -1,8 +1,15 @@
 import { Button, Col, Form, Input, Row } from 'antd'
-import { getConfig, saveConfig } from 'providers/configProvider'
 import React, { useEffect } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { getConfig, saveConfig } from '../providers/configProvider'
 
 export default function Login() {
+  const requireLogin = getConfig().server === '' || getConfig().privateKey === ''
+  const navigate = useNavigate()
+  useEffect(() => {
+    if (!requireLogin) navigate('/')
+  }, [])
+
   const [form] = Form.useForm()
 
   useEffect(() => {
@@ -11,7 +18,7 @@ export default function Login() {
 
   const login = async () => {
     saveConfig(form.getFieldsValue())
-    location.href = '/'
+    navigate('/', { replace: true }) // 不在 history 里留记录，因为不应该留
   }
 
   return (
